requirements.txt
contourpy==1.1.1
cycler==0.12.1
filelock==3.13.4
fonttools==4.51.0
fsspec==2024.3.1
imageio==2.34.1
imporesources==6.4.0
jinja2==3.1.3
kiwisolver==1.4.5
oader==0.4
markupsafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
munch==4.0.0
networkx==3.1
numpy==1.24.4
packaging==24.0
patchify==0.2.3
pillow==10.3.0
pyparsing==3.1.2
pyateutil==2.9.0.post0
pywavelets==1.4.1
scmage==0.21.0
scipy==1.10.1
six==1.16.0
sympy==1.12
tifffile==2023.7.10
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.2
tyxtensions==4.11.0
zipp==3.18.1
Pyyaml